backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.2
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
psycopg2